# server/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "七日记（Note4Seven）API",
    "version": "0.1.0",
    "description": "Generated route inventory with shared runtime contracts. Examples must use synthetic data."
  },
  "servers": [
    {
      "url": "https://staging.invalid",
      "description": "Staging placeholder; inject the real domain outside source control"
    }
  ],
  "paths": {
...
    "/api/v1/home/modules": {
      "get": {
        "operationId": "get_api_v1_home_modules",
        "tags": [
          "v1"
        ],
        "summary": "GET /api/v1/home/modules",
        "x-source": "src/routes/modules.ts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/module-templates": {
      "get": {
        "operationId": "get_api_v1_module_templates",
        "tags": [
          "v1"
        ],
        "summary": "GET /api/v1/module-templates",
        "x-source": "src/routes/modules.ts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/modules": {
      "post": {
        "operationId": "post_api_v1_modules",
        "tags": [
          "v1"
        ],
        "summary": "POST /api/v1/modules",
        "x-source": "src/routes/modules.ts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/modules/{moduleId}": {
      "get": {
        "operationId": "get_api_v1_modules_moduleId",
        "tags": [
          "v1"
        ],
        "summary": "GET /api/v1/modules/{moduleId}",
        "x-source": "src/routes/modules.ts",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "patch": {
        "operationId": "patch_api_v1_modules_moduleId",
        "tags": [
          "v1"
        ],
        "summary": "PATCH /api/v1/modules/{moduleId}",
        "x-source": "src/routes/modules.ts",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/modules/{moduleId}/delete": {
      "post": {
        "operationId": "post_api_v1_modules_moduleId_delete",
        "tags": [
          "v1"
        ],
        "summary": "POST /api/v1/modules/{moduleId}/delete",
        "x-source": "src/routes/modules.ts",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/modules/{moduleId}/pin": {
      "put": {
        "operationId": "put_api_v1_modules_moduleId_pin",
        "tags": [
          "v1"
        ],
        "summary": "PUT /api/v1/modules/{moduleId}/pin",
        "x-source": "src/routes/modules.ts",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/modules/{moduleId}/restore": {
      "post": {
        "operationId": "post_api_v1_modules_moduleId_restore",
        "tags": [
          "v1"
        ],
        "summary": "POST /api/v1/modules/{moduleId}/restore",
        "x-source": "src/routes/modules.ts",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/modules/{moduleId}/settings": {
      "get": {
        "operationId": "get_api_v1_modules_moduleId_settings",
        "tags": [
          "v1"
        ],
        "summary": "GET /api/v1/modules/{moduleId}/settings",
        "x-source": "src/routes/modules.ts",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/modules/recycle-bin": {
      "get": {
        "operationId": "get_api_v1_modules_recycle_bin",
        "tags": [
          "v1"
        ],
        "summary": "GET /api/v1/modules/recycle-bin",
        "x-source": "src/routes/modules.ts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "Error": {
        "description": "Standard error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Quota exceeded",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Timestamp": {
        "type": "string",
        "format": "date-time"
      },
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "requestId",
          "serverTime"
        ],
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": "OK"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "requestId": {
            "type": "string"
          },
          "serverTime": {
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId",
          "serverTime"
        ],
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": [
              "object",
              "null"
            ]
          },
          "requestId": {
            "type": "string"
          },
          "serverTime": {
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
...
      "ModuleCreateRequest": {
        "type": "object",
        "required": [
          "name",
          "recordPolicy",
          "clientRequestId"
        ],
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10
          },
          "description": {
            "type": "string",
            "maxLength": 200,
            "default": ""
          },
          "recordPolicy": {
            "type": "string",
            "enum": [
              "strict",
              "relaxed"
            ],
            "description": "Immutable after module creation."
          },
          "templateId": {
            "type": "string",
            "maxLength": 64
          },
          "clientRequestId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 64
          }
        }
      },
...
    }
  }
}